refactor pure_with
